[GH-6] Refactor parser API and structure.
Create three parsers: ZIP that decorates another parser, EA-compatible that transforms input before processing it, and EMF-based one.
