+++ src/test/java/com/akaene/stpa/scs/parser/sysml/EnterpriseArchitectSysMLXMIParserTest.java
@@ -4,19 +4,20 @@ import org.junit.jupiter.api.Test;
 
 import java.io.File;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class EnterpriseArchitectTransformerTest {
+class EnterpriseArchitectSysMLXMIParserTest {
 
     @Test
     void isEnterpriseArchitectFileReturnsTrueForEA() throws Exception {
         final File input = new File(getClass().getClassLoader().getResource("simple-model_EA.xml").toURI());
-        assertTrue(EnterpriseArchitectTransformer.isEnterpriseArchitectFile(input));
+        assertTrue(EnterpriseArchitectSysMLXMIParser.isEnterpriseArchitectFile(input));
     }
 
     @Test
     void isEnterpriseArchitectFileReturnsFalseForNonEA() throws Exception {
         final File input = new File(getClass().getClassLoader().getResource("simple-model/model.xmi").toURI());
-        assertFalse(EnterpriseArchitectTransformer.isEnterpriseArchitectFile(input));
+        assertFalse(EnterpriseArchitectSysMLXMIParser.isEnterpriseArchitectFile(input));
     }
 }
